This commit contains just the gitignore file.
This makes it useful as a starting point for existing repositories.

E.G. It is possible to add a remote to this repo and create a new branch from this commit.  After adding the repo's files and committing, merging in subsequent commits from one of the other branches can be done with minimal conflict.